fix: normalize thresholding; fix parameters

+++ DifferenceOfGaussians/Lib/FilterSettings.cs
@@ -15,7 +15,7 @@ namespace DifferenceOfGaussians.Lib
 
     public class ThresholdSettings
     {
-        public int ThresholdValue { get; set; } = 128;
+        public double ThresholdValue { get; set; } = 128;
         public double Phi { get; set; } = 1.0;
     }
 }
